Get ACE, core and coupled tests passing on MPS (#1672)
I fixed/skipped all relevant tests in `fme/core`, `fme/ace`, and
`fme/coupled` when calling `FME_USE_MPS=1 pytest`.
I didn't touch `fme/downscaling` since there is a lot in flux there
right now.

+++ fme/ace/aggregator/inference/test_inference.py
@@ -65,7 +65,7 @@ def test_logs_labels_exist_with_reference_time_means():
     reference_time_means = xr.Dataset(
         {
             "a": xr.DataArray(
-                np.random.randn(ny, nx),
+                np.random.randn(ny, nx).astype(np.float32),
                 dims=["grid_yt", "grid_xt"],
             )
         }
